backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pdfplumber==0.11.4
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.9.0
